# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LeadFlow AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business:
    get:
      operationId: getBusiness
      tags: [business]
      summary: Get the current business
      responses:
        "200":
          description: Business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    patch:
      operationId: updateBusiness
      tags: [business]
      summary: Update the current business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads for the current business
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Lead list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadList"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create and qualify a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{leadId}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead with its conversation and scoring reasons
      parameters:
        - $ref: "#/components/parameters/LeadId"
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetail"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead status
      parameters:
        - $ref: "#/components/parameters/LeadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - $ref: "#/components/parameters/LeadId"
      responses:
        "204":
          description: Deleted
...
  /conversations/{conversationId}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a conversation with messages
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "200":
          description: Conversation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
    patch:
      operationId: updateConversation
      tags: [conversations]
      summary: Update conversation status
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationUpdate"
      responses:
        "200":
          description: Updated conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /assistant:
    get:
      operationId: getAssistant
      tags: [assistant]
      summary: Get assistant configuration
      responses:
        "200":
          description: Assistant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
    patch:
      operationId: updateAssistant
      tags: [assistant]
      summary: Update assistant configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantUpdate"
      responses:
        "200":
          description: Updated assistant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
  /automations:
    get:
      operationId: listAutomations
      tags: [automation]
      summary: Get automation settings
      responses:
        "200":
          description: Automations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    patch:
      operationId: updateAutomation
      tags: [automation]
      summary: Update an automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationUpdate"
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
...
  /chat/{businessSlug}:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to a business chatbot
      parameters:
        - in: path
          name: businessSlug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
components:
  parameters:
    LeadId:
      in: path
      name: leadId
      required: true
      schema:
        type: string
    ConversationId:
      in: path
      name: conversationId
      required: true
      schema:
        type: string
    Search:
      in: query
      name: search
      required: false
      schema:
        type: string
    Status:
      in: query
      name: status
      required: false
      schema:
        type: string
    Page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        default: 1
    PageSize:
      in: query
      name: pageSize
      required: false
      schema:
        type: integer
        default: 20
  schemas:
...
    Business:
      type: object
      required: [id, name, slug, industry, description]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        industry: { type: string }
        description: { type: string }
        location: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
    BusinessUpdate:
      type: object
      properties:
        name: { type: string }
        industry: { type: string }
        description: { type: string }
        location: { type: string }
        website: { type: string }
...
    Lead:
      type: object
      required: [id, name, email, phone, service, budget, location, timeline, score, status, source, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        service: { type: string }
        budget: { type: string }
        location: { type: string }
        timeline: { type: string }
        score: { type: integer, minimum: 0, maximum: 100 }
        status: { type: string }
        source: { type: string }
        createdAt: { type: string }
        conversationId: { type: ["string", "null"] }
    LeadList:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Lead" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    LeadInput:
      type: object
      required: [name, email, phone, service, budget, location, timeline, source]
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        service: { type: string }
        budget: { type: string }
        location: { type: string }
        timeline: { type: string }
        source: { type: string }
        conversationId: { type: string }
    LeadUpdate:
      type: object
      properties:
        status: { type: string }
    QualificationReason:
      type: object
      required: [label, points]
      properties:
        label: { type: string }
        points: { type: integer }
    LeadDetail:
      allOf:
        - $ref: "#/components/schemas/Lead"
        - type: object
          required: [qualificationReasons, conversation]
          properties:
            qualificationReasons:
              type: array
              items: { $ref: "#/components/schemas/QualificationReason" }
            conversation:
              $ref: "#/components/schemas/ConversationDetail"
    Message:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id: { type: string }
        role: { type: string }
        content: { type: string }
        createdAt: { type: string }
    Conversation:
      type: object
      required: [id, visitorName, status, startedAt, lastActivityAt, messageCount]
      properties:
        id: { type: string }
        visitorName: { type: string }
        status: { type: string }
        startedAt: { type: string }
        lastActivityAt: { type: string }
        messageCount: { type: integer }
        leadId: { type: ["string", "null"] }
    ConversationDetail:
      allOf:
        - $ref: "#/components/schemas/Conversation"
        - type: object
          required: [messages]
          properties:
            messages:
              type: array
              items: { $ref: "#/components/schemas/Message" }
            lead:
              anyOf:
                - $ref: "#/components/schemas/Lead"
                - type: "null"
    ConversationUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
    FAQ:
      type: object
      required: [question, answer]
      properties:
        question: { type: string }
        answer: { type: string }
    Assistant:
      type: object
      required: [id, name, greeting, tone, description, instructions, services, faqs, openingHours, locations]
      properties:
        id: { type: string }
        name: { type: string }
        greeting: { type: string }
        tone: { type: string }
        description: { type: string }
        instructions: { type: string }
        services:
          type: array
          items: { type: string }
        faqs:
          type: array
          items: { $ref: "#/components/schemas/FAQ" }
        openingHours: { type: string }
        locations: { type: string }
    AssistantUpdate:
      type: object
      properties:
        name: { type: string }
        greeting: { type: string }
        tone: { type: string }
        description: { type: string }
        instructions: { type: string }
        services:
          type: array
          items: { type: string }
        faqs:
          type: array
          items: { $ref: "#/components/schemas/FAQ" }
        openingHours: { type: string }
        locations: { type: string }
    Automation:
      type: object
      required: [id, key, label, description, enabled]
      properties:
        id: { type: string }
        key: { type: string }
        label: { type: string }
        description: { type: string }
        enabled: { type: boolean }
    AutomationUpdate:
      type: object
      required: [key, enabled]
      properties:
        key: { type: string }
        enabled: { type: boolean }
...
    ChatInput:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1 }
        conversationId: { type: ["string", "null"] }
    ChatResponse:
      type: object
      required: [conversationId, message, assistantName, leadCaptured]
      properties:
        conversationId: { type: string }
        message: { type: string }
        assistantName: { type: string }
        leadCaptured: { type: boolean }
        leadId: { type: ["string", "null"] }