# lib/api-zod/src/generated/types/assistantUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LeadFlow AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Faq } from './faq';

export interface AssistantUpdate {
  name?: string;
  greeting?: string;
  tone?: string;
  description?: string;
  instructions?: string;
  services?: string[];
  faqs?: Faq[];
  openingHours?: string;
  locations?: string;
}