# lib/api-zod/src/generated/types/automation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LeadFlow AI API
 * OpenAPI spec version: 0.1.0
 */

export interface Automation {
  id: string;
  key: string;
  label: string;
  description: string;
  enabled: boolean;
}