# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LeadFlow AI API
 * OpenAPI spec version: 0.1.0
 */

export interface Business {
  id: string;
  name: string;
  slug: string;
  industry: string;
  description: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  website?: string | null;
}